requirements.txt
discord.py
pyshorteners
pynvml